bb-mesh-service/requirements.txt
paho-mqtt>=1.6.1,<2.0.0
meshtastic>=2.3.11,<3.0.0
google-cloud-datastore>=2.19.0,<3.0.0
cryptography>=41.0.7,<42.0.0
protobuf>=5.26.0,<6.0.0
flask>=3.0.0,<4.0.0
gunicorn>=21.2.0,<22.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
structlog>=23.2.0,<24.0.0
requests>=2.31.0,<3.0.0
